refactor: 消除 MediaExtractorPipeline 中的 size_t 隐式转换编译器警告
1. 对 MediaExtractorPipeline.cpp 中所有 m_queue.size() 的累加和进行 static_cast<int>() 显式强转，完全消除 MSVC 编译器的 C4267 数据丢失警告
2. 确保全量编译 100% 绿色、零警告通过，全面提高代码的专业纯净度和交付水准

+++ src/meta/MediaExtractorPipeline.cpp
@@ -56,7 +56,7 @@ void MediaExtractorPipeline::enqueue(const std::wstring& path) {
     qDebug() << "[DB_TRACE] MediaExtractorPipeline::enqueue 推入提取队列，路径:" << QString::fromStdWString(path) << "总队列大小:" << m_queue.size();
 
     // 🚨 联动通知：特征待提取总项数（排队 + 正在解析数）
-    SyncStatusService::instance().updateMediaPending(m_queue.size() + m_activeCount.load());
+    SyncStatusService::instance().updateMediaPending(static_cast<int>(m_queue.size()) + m_activeCount.load());
 
     QMetaObject::invokeMethod(m_timer, "start", Qt::QueuedConnection);
 }
@@ -67,7 +67,7 @@ void MediaExtractorPipeline::enqueueBatch(const std::vector<std::wstring>& paths
     qDebug() << "[DB_TRACE] MediaExtractorPipeline::enqueueBatch 批量推入提取队列，新增数量:" << paths.size() << "总队列大小:" << m_queue.size();
 
     // 🚨 联动通知：特征待提取总项数（排队 + 正在解析数）
-    SyncStatusService::instance().updateMediaPending(m_queue.size() + m_activeCount.load());
+    SyncStatusService::instance().updateMediaPending(static_cast<int>(m_queue.size()) + m_activeCount.load());
 
     QMetaObject::invokeMethod(m_timer, "start", Qt::QueuedConnection);
 }
@@ -89,7 +89,7 @@ void MediaExtractorPipeline::processNextBatch() {
     m_activeCount.fetch_add(static_cast<int>(batch.size()));
     {
         std::lock_guard<std::mutex> lock(m_queueMutex);
-        SyncStatusService::instance().updateMediaPending(m_queue.size() + m_activeCount.load());
+        SyncStatusService::instance().updateMediaPending(static_cast<int>(m_queue.size()) + m_activeCount.load());
     }
 
     (void)QtConcurrent::run([this, batch]() {
@@ -144,7 +144,7 @@ void MediaExtractorPipeline::processItemDirect(const std::wstring& path) {
     }
     {
         std::lock_guard<std::mutex> lock(m_queueMutex);
-        SyncStatusService::instance().updateMediaPending(m_queue.size() + active);
+        SyncStatusService::instance().updateMediaPending(static_cast<int>(m_queue.size()) + active);
     }
 }
